requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-community
langchain-openai
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
pydantic
